Friend list information is now stored on the server on user log-in.
Client periodically polls /downloadFriendList to keep in touch with server register updates.

IMPORTANT:

Application is not yet configured with Webhooks to Facebook's API.
This implies that live changes to friend lists in Facebook are not reflected in Pause because of API data caps.

Webhooks configuration requires an HTTPS server with a valid TCL/SSL certificate, hence a necessary future server migration to Heroku.

Fb Webhooks Doc:
https://developers.facebook.com/docs/graph-api/webhooks/